1. Parcel Endpoint Created.

See merge request rdamyanova/deliverit13!6

+++ DeliverIt13/DeliverIt13.Web/Controllers/WarehouseController.cs
@@ -1,7 +1,8 @@
 ﻿using System;
-using DeliverIt13.Data.Models;
+using DeliverIt13.Data.Enums;
 using DeliverIt13.Services.Contracts;
 using DeliverIt13.Services.Models;
+using DeliverIt13.Web.Helpers;
 using Microsoft.AspNetCore.Mvc;
 
 namespace DeliverIt13.Web.Controllers
@@ -11,18 +12,25 @@ namespace DeliverIt13.Web.Controllers
     public class WarehouseController : ControllerBase
     {
         private readonly IWarehouseService warehouseService;
+        private readonly IAuthHelper authHelper;
 
-        public WarehouseController(IWarehouseService warehouseService)
+        public WarehouseController(IWarehouseService warehouseService,IAuthHelper authHelper)
         {
             this.warehouseService = warehouseService;
+            this.authHelper = authHelper;
         }
 
 
         [HttpGet("{id}")]
-        public IActionResult Get(int id)
+        public IActionResult Get(int id,[FromHeader] string credentials)
         {
             try
             {
+                var user = this.authHelper.TryGetUser(credentials);
+                if (user.Type != UserType.Employee)
+                {
+                    return Unauthorized($"Unauthorized Username/Email: {credentials}");
+                }
                 var warehouse = this.warehouseService.Get(id);
                 return Ok(warehouse);
             }
@@ -32,11 +40,16 @@ namespace DeliverIt13.Web.Controllers
             }
         }
 
-        [HttpGet("")]
-        public IActionResult GetAll()
+        [HttpGet("internal")]
+        public IActionResult GetAll([FromHeader] string credentials)
         {
             try
             {
+                var user = this.authHelper.TryGetUser(credentials);
+                if (user.Type != UserType.Employee)
+                {
+                    return Unauthorized($"Unauthorized Username/Email: {credentials}");
+                }
                 var warehouses = this.warehouseService.GetAll();
                 return Ok(warehouses);
             }
@@ -46,12 +59,31 @@ namespace DeliverIt13.Web.Controllers
             }
         }
 
+        [HttpGet("")]
+        public IActionResult GetAllPublic()
+        {
+            try
+            {
+                var warehouses = this.warehouseService.GetAllPublic();
+                return Ok(warehouses);
+            }
+            catch (Exception e)
+            {
+                return BadRequest(e.Message);
+            }
+        }
+
         [HttpPost("")]
-        public IActionResult Post([FromBody] WarehouseDTO warehouse)
+        public IActionResult Post([FromBody] WarehouseCreateDTO warehouse, [FromHeader] string credentials)
         {
 
             try
             {
+                var user = this.authHelper.TryGetUser(credentials);
+                if (user.Type != UserType.Employee)
+                {
+                    return Unauthorized($"Unauthorized Username/Email: {credentials}");
+                }
                 this.warehouseService.Create(warehouse);
                 return Created("Warehouse Created", warehouse);
             }
@@ -62,10 +94,15 @@ namespace DeliverIt13.Web.Controllers
         }
 
         [HttpDelete("{id}")]
-        public IActionResult Delete(int id)
+        public IActionResult Delete(int id,[FromHeader] string credentials)
         {
             try
             {
+                var user = this.authHelper.TryGetUser(credentials);
+                if (user.Type != UserType.Employee)
+                {
+                    return Unauthorized($"Unauthorized Username/Email: {credentials}");
+                }
                 this.warehouseService.Delete(id);
                 return NoContent();
             }
@@ -76,10 +113,15 @@ namespace DeliverIt13.Web.Controllers
         }
 
         [HttpPut("")]
-        public IActionResult Put([FromBody] WarehouseUpdateDTO warehouse)
+        public IActionResult Put([FromBody] WarehouseUpdateDTO warehouse,[FromHeader] string credentials)
         {
             try
             {
+                var user = this.authHelper.TryGetUser(credentials);
+                if (user.Type != UserType.Employee)
+                {
+                    return Unauthorized($"Unauthorized Username/Email: {credentials}");
+                }
                 var updatedWarehouse = this.warehouseService.Update(warehouse);
                 return Ok(updatedWarehouse);
             }
